third_party/WebKit/LayoutTests/editing/inserting/insert-div-021.html

<!doctype html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../assert_selection.js"></script>
<div id="log"></div>
<script>
test(() => assert_selection(
    [
        '<div contenteditable>',
            '<div style="height: 12px"></div>',
            'foo<br>|bar',
        '</div>',
    ].join(''),
    selection => {
        selection.document.execCommand('delete');
        selection.document.execCommand('insertParagraph');
        selection.document.execCommand('insertParagraph');
    },
    [
        '<div contenteditable>',
            '<div style="height: 12px"></div>',
            'foo<div><br></div>',
            '<div>|bar</div>',
        '</div>',
    ].join('')),
    'Insert visible newline after deleting');
</script>
